Refactor: cambio de decimal a int

+++ creates_schema.sql
@@ -12,9 +12,9 @@ CREATE TABLE ventas (
     producto VARCHAR(100),
     referencia VARCHAR(50),
     cantidad INT NOT NULL,
-    valor_unitario DECIMAL(15,2) NOT NULL,
-    valor_vendido DECIMAL(15,2) NOT NULL,
-    impuesto DECIMAL(15,2) DEFAULT 0,
+    valor_unitario INT NOT NULL,
+    valor_vendido INT NOT NULL,
+    impuesto INT DEFAULT 0,
     FOREIGN KEY (vendedor_id) REFERENCES vendedores(vendedor_id)
 );
 CREATE TABLE devoluciones (
@@ -24,9 +24,9 @@ CREATE TABLE devoluciones (
     producto VARCHAR(100),
     referencia VARCHAR(50),
     cantidad INT NOT NULL,
-    valor_unitario DECIMAL(15,2) NOT NULL,
-    valor_vendido DECIMAL(15,2) NOT NULL,
-    impuesto DECIMAL(15,2) DEFAULT 0,
+    valor_unitario INT NOT NULL,
+    valor_vendido INT NOT NULL,
+    impuesto INT DEFAULT 0,
     tipo_operacion VARCHAR(50),
     motivo TEXT,
     FOREIGN KEY (vendedor_id) REFERENCES vendedores(vendedor_id)
@@ -36,13 +36,13 @@ CREATE TABLE comisiones (
     vendedor_id INT NOT NULL,
     anio INT NOT NULL,
     mes INT NOT NULL,
-    total_ventas DECIMAL(15,2) NOT NULL,
-    total_devoluciones DECIMAL(15,2) NOT NULL,
+    total_ventas INT NOT NULL,
+    total_devoluciones INT NOT NULL,
     porcentaje_base DECIMAL(5,2) NOT NULL,       -- 5.00 por el 5%
     bono DECIMAL(5,2) DEFAULT 0,                 -- 2.00 por el 2%
     penalizacion DECIMAL(5,2) DEFAULT 0,         -- -1.00 por el -1%
     porcentaje_final DECIMAL(5,2) NOT NULL,
-    valor_comision DECIMAL(15,2) NOT NULL,
+    valor_comision INT NOT NULL,
     fecha_calculo TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (vendedor_id) REFERENCES vendedores(vendedor_id)
 );
@@ -50,5 +50,5 @@ CREATE TABLE productos (
     producto_id INT AUTO_INCREMENT PRIMARY KEY,
     nombre VARCHAR(100) NOT NULL,
     referencia VARCHAR(50),
-    precio_base DECIMAL(15,2)
+    precio_base INT
 );
